+++ cmd/errtrace/main.go
@@ -32,6 +32,7 @@ package main
 import (
 	"bytes"
 	"encoding/json"
+	"errors"
 	"flag"
 	"fmt"
 	"go/ast"
@@ -47,6 +48,7 @@ import (
 	"sort"
 	"strconv"
 	"strings"
+	"time"
 )
 
 func main() {
@@ -63,6 +65,8 @@ type mainParams struct {
 	List     bool     // -l
 	Format   format   // -format
 	Patterns []string // list of files to process
+
+	ImplicitStdin bool // whether stdin was picked because there were no args
 }
 
 func (p *mainParams) shouldFormat() bool {
@@ -103,6 +107,7 @@ func (p *mainParams) Parse(w io.Writer, args []string) error {
 	if len(p.Patterns) == 0 {
 		// Read file from stdin when there's no args, similar to gofmt.
 		p.Patterns = []string{"-"}
+		p.ImplicitStdin = true
 	}
 
 	return nil
@@ -208,11 +213,12 @@ func (cmd *mainCmd) Run(args []string) (exitCode int) {
 		}
 
 		req := fileRequest{
-			Format:   p.shouldFormat(),
-			Write:    p.Write,
-			List:     p.List,
-			Filename: display,
-			Filepath: file,
+			Format:        p.shouldFormat(),
+			Write:         p.Write,
+			List:          p.List,
+			Filename:      display,
+			Filepath:      file,
+			ImplicitStdin: p.ImplicitStdin,
 		}
 		if err := cmd.processFile(req); err != nil {
 			cmd.log.Printf("%s:%s", display, err)
@@ -323,6 +329,8 @@ type fileRequest struct {
 
 	Filename string // name displayed to the user
 	Filepath string // actual location on disk, or "-" for stdin
+
+	ImplicitStdin bool
 }
 
 // processFile processes a single file.
@@ -550,15 +558,50 @@ func (cmd *mainCmd) processFile(r fileRequest) error {
 	return err
 }
 
+var _stdinWait = 200 * time.Millisecond
+
 func (cmd *mainCmd) readFile(r fileRequest) ([]byte, error) {
-	if r.Filepath == "-" {
-		if r.Write {
-			return nil, fmt.Errorf("can't use -w with stdin")
-		}
+	if r.Filepath != "-" {
+		return os.ReadFile(r.Filename)
+	}
+
+	if r.Write {
+		return nil, fmt.Errorf("can't use -w with stdin")
+	}
+
+	if !r.ImplicitStdin {
 		return io.ReadAll(cmd.Stdin)
 	}
 
-	return os.ReadFile(r.Filepath)
+	// If we're reading from stdin because there were no other arguments,
+	// wait a short time for the first read.
+	// If there's nothing, print a warning and continue waiting.
+	firstRead := make(chan struct{})
+	go func(firstRead <-chan struct{}) {
+		select {
+		case <-firstRead:
+		case <-time.After(_stdinWait):
+			cmd.log.Println("reading from stdin; use '-h' for help")
+		}
+	}(firstRead)
+
+	var buff bytes.Buffer
+	bs := make([]byte, 1024)
+	for {
+		n, err := cmd.Stdin.Read(bs)
+		buff.Write(bs[:n])
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				err = nil
+			}
+			return buff.Bytes(), err
+		}
+
+		if firstRead != nil {
+			close(firstRead)
+			firstRead = nil
+		}
+	}
 }
 
 type walker struct {