+++ cmd/errtrace/main_test.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"go/parser"
 	"go/token"
+	"io"
 	"os"
 	"os/exec"
 	"path"
@@ -16,6 +17,7 @@ import (
 	"strings"
 	"sync"
 	"testing"
+	"time"
 
 	"braces.dev/errtrace/internal/diff"
 )
@@ -206,7 +208,8 @@ func TestParseMainParams(t *testing.T) {
 		{
 			name: "stdin",
 			want: mainParams{
-				Patterns: []string{"-"},
+				Patterns:      []string{"-"},
+				ImplicitStdin: true,
 			},
 		},
 	}
@@ -449,7 +452,7 @@ func TestFormatAuto(t *testing.T) {
 			Stdin:  strings.NewReader("unused"),
 			Stdout: &out,
 			Stderr: &err,
-		}).Run([]string{"-w"})
+		}).Run([]string{"-w", "-"})
 		if want := 1; exitCode != want {
 			t.Errorf("exit code = %d, want %d", exitCode, want)
 		}
@@ -711,6 +714,73 @@ func TestGoListFilesBadJSON(t *testing.T) {
 
 	if want, got := "go list: output malformed: invalid character 'b'", stderr.String(); !strings.Contains(got, want) {
 		t.Errorf("stderr = %q, want %q", got, want)
+	}
+}
+
+func TestStdinNoInputMessage(t *testing.T) {
+	// Verify that if there's no input on implicit stdin,
+	// we print a message to stderr to help the user.
+	defer func(old time.Duration) { _stdinWait = old }(_stdinWait)
+	_stdinWait = 10 * time.Millisecond // make the test run faster
+
+	tests := []struct {
+		name       string
+		delay      time.Duration // before writing
+		args       []string
+		wantStderr string
+	}{
+		{
+			name:  "no delay",
+			delay: 0,
+		},
+		{
+			name:       "delay",
+			delay:      20 * time.Millisecond,
+			wantStderr: "reading from stdin; use '-h' for help\n",
+		},
+		{
+			name:  "explicit stdin with delay",
+			args:  []string{"-"},
+			delay: 20 * time.Millisecond,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// mainCmd.Run is blocking so this has to be in a goroutine.
+			stdin, stdinw := io.Pipe()
+			done := make(chan struct{})
+			go func() {
+				defer close(done)
+
+				if tt.delay > 0 {
+					time.Sleep(20 * time.Millisecond)
+				}
+				if _, err := io.WriteString(stdinw, "package foo\n"); err != nil {
+					t.Error(err)
+				}
+
+				if err := stdinw.Close(); err != nil {
+					t.Error(err)
+				}
+			}()
+			defer func() { <-done }()
+
+			var stderr bytes.Buffer
+			exitCode := (&mainCmd{
+				Stdin:  stdin,
+				Stdout: io.Discard,
+				Stderr: &stderr,
+			}).Run(tt.args)
+
+			if want := 0; exitCode != want {
+				t.Errorf("exit code = %d, want %d", exitCode, want)
+			}
+
+			if want, got := tt.wantStderr, stderr.String(); got != want {
+				t.Errorf("stderr = %q, want %q", got, want)
+			}
+		})
 	}
 }